+++ changelog.md
@@ -2,6 +2,7 @@
 
 - 10:30 - Backend: refactor video download-related code and comments. Used Celery's built-in tools for retrying 
   failed tasks.
+- 14:00 - Frontend: refactored key-related code for better reactive UI updates.
 
 ## 23.04.2023, Sun